Part1/requirements.txt
streamlit
azure-ai-formrecognizer
openai
